backend/pyproject.toml
[project]
name = "todo-backend-api"
version = "0.1.0"
description = "FastAPI backend for multi-user Todo application with JWT auth"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "pyjwt[crypto]>=2.9.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "httpx>=0.27.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "cryptography>=42.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["app"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
